+++ plugin/src/main/kotlin/de/infolektuell/gradle/typst/extensions/TypstSourceSet.kt
@@ -1,20 +1,30 @@
 package de.infolektuell.gradle.typst.extensions
 
+import org.gradle.api.Action
 import org.gradle.api.Named
 import org.gradle.api.file.Directory
 import org.gradle.api.file.DirectoryProperty
-import org.gradle.api.provider.*
+import org.gradle.api.provider.ListProperty
+import org.gradle.api.provider.MapProperty
+import org.gradle.api.provider.Provider
+import org.gradle.api.provider.SetProperty
+import org.gradle.api.tasks.Nested
 
 abstract class TypstSourceSet : Named {
     abstract val destinationDir: DirectoryProperty
     abstract val documents: ListProperty<String>
     abstract val inputs: MapProperty<String, String>
-    abstract val merged: Property<String>
 
     abstract val typst: SetProperty<Directory>
     abstract val data: SetProperty<Directory>
     abstract val images: SetProperty<Directory>
     abstract val fonts: SetProperty<Directory>
+    @get:Nested
+    abstract val format: TypstOutputFormatExtension
+    fun format(action: Action<in TypstOutputFormatExtension>) {
+        action.execute(format)
+    }
+
   fun addSourceSet(sourceSet: TypstSourceSet) {
     typst.addAll(sourceSet.typst)
     data.addAll(sourceSet.data)